it works!

+++ Quake/gl_rmain.c
@@ -474,8 +474,8 @@ float frustum_skew = 0.0; //used by r_stereo
 void GL_SetFrustum(float fovx, float fovy)
 {
 	float xmax, ymax;
-	xmax = NEARCLIP * tan( fovx * M_PI / 360.0 );
-	ymax = NEARCLIP * tan( fovy * M_PI / 360.0 );
+	xmax = NEARCLIP * tan( (double)fovx * M_PI / 360.0 );
+	ymax = NEARCLIP * tan( (double)fovy * M_PI / 360.0 );
 	glFrustum(-xmax + frustum_skew, xmax + frustum_skew, -ymax, ymax, NEARCLIP, gl_farclip.value);
 }
